AI_Triage/backend/requirements.txt

Flask
flask-cors
flask-mysqldb
pandas
joblib
scikit-learn
mysql-connector-python
opencv-python
scipy
deepface
tf-keras
